wayland: Correctly compare refresh rates when applying config
KScreen refresh rate is in Hz, the wayland protocol uses mHz. In order
to address the issue, this change makes the WaylandOutputDeviceMode
provide the refresh rate in the same unit as KScreen::Mode, i.e. Hz.

+++ backends/kwayland/waylandoutputdevice.h
@@ -38,7 +38,7 @@ public:
     qreal scale() const;
     QPoint globalPosition() const;
     QSize pixelSize() const;
-    int refreshRate() const;
+    float refreshRate() const;
     uint32_t vrrPolicy() const;
     uint32_t overscan() const;
     uint32_t capabilities() const;
